+++ lib/src/file_downloader.dart
@@ -1089,10 +1089,11 @@ interface class FileDownloader {
   /// Restarts running upload tasks if [alsoRestartUploads] is true (which
   /// requires [rescheduleRunningTasks] to be true as well).
   ///
-  /// Note that on Android, requiring WiFi enforces an unmetered network
-  /// connection constraint (`NETWORK_TYPE_UNMETERED`), excluding metered Wi-Fi
-  /// hotspots and cellular data. On iOS, requiring WiFi disables cellular access
-  /// (`allowsCellularAccess = false`), allowing execution over any Wi-Fi or Ethernet.
+  /// Note that on Android (API 28+), requiring WiFi enforces a Wi-Fi network constraint
+  /// (`TRANSPORT_WIFI`), preventing tasks from running over cellular connections even if unmetered.
+  /// On older Android versions, it enforces an unmetered network constraint (`NETWORK_TYPE_UNMETERED`).
+  /// On iOS, requiring WiFi disables cellular access (`allowsCellularAccess = false`), allowing
+  /// execution over any Wi-Fi or Ethernet connection.
   Future<bool> requireWiFi(
     RequireWiFi requirement, {
     rescheduleRunningTasks = true,